Added Firebase Analytics
This commit adds the firebase analytics package file giving all requirements for this, it should compile fine but you must have both android and iOS builds installed to make this work.
Should also be conjoined with the package name, but should be checked to ensure that it's present in settings (will check over myself)

Any issues please add comments to this commit.