Added Data Peristence
-Data persists for both changes in scene and for changes in session
-Implemented using JSON Serializability

MainManager is an example of a persisting class that is attached to an
empty object of the same name.

Awake() is called on when the object the
script is attached to is initialised or set to active. The current Awake
method confirms the class is singleton across all scenes using a stored
instance of MainManager by using the unity method DontDestroyOnLoad().

The Class SaveData stores the data to be saved, marke by Serializable to
allow JSON parsing.

SaveSettings() stores the SaveData variables in a JSON file,
Application.persistentDataPath is a different location on different
systems.

LoadSettings() gets the path and if exists it fetches the data and
stores into the according SaveData variables.